feat: fact_sales with CTEs for obtaining keys from dimensions
the incremental loading is not functioning correctly, due to the lack of
product history. Since SCD 2 for dim_product is implemented after 2014,
all orders placed before that time will be lost after the dim_product
join logic. Can only be fixed by manually changing the ModifiedDate and
SCD time of the Product-related snapshots.